requirements.txt
Django==1.10.3
django-allauth==0.29.0
mysqlclient==1.3.9
Pillow==3.4.2

defusedxml==0.4.1
oauthlib==2.0.1
python3-openid==3.0.10
requests==2.12.1
requests-oauthlib==0.7.0